+++ db4e/Panes/P2PoolPane.py
@@ -58,7 +58,7 @@ class P2PoolPane(Container):
 
     health_msgs = Label()
 
-    analytics_button = Button(label=DLabel.ANALYTICS, id=DButton.ANALYTICS)
+    shares_found_button = Button(label=DLabel.SHARES_FOUND, id=DButton.SHARES_FOUND)
     delete_button = Button(label=DLabel.DELETE, id=DButton.DELETE)
     disable_button = Button(label=DLabel.STOP, id=DButton.DISABLE)
     enable_button = Button(label=DLabel.START, id=DButton.ENABLE)
@@ -111,8 +111,8 @@ class P2PoolPane(Container):
 
                 Vertical(
                     Horizontal(
-                        self.analytics_button,
                         self.hashrate_button,
+                        self.shares_found_button,
                         self.new_button,
                         self.update_button,
                         self.enable_button,