+++ db4e/data_migration.py
@@ -9,6 +9,29 @@ from datetime import datetime, timedelta
 db = DbMgr()
 
 
+def modify_block_found_events_2():
+    recs = db.find_many(col_name="mining", filter={"doc_type": "block_found_event"})
+    for rec in recs:
+        new_rec = {
+            DMongo.DOC_TYPE: "block_found_event",
+            DMongo.CHAIN: "minisidechain",
+            DMongo.POOL: rec[DMongo.INSTANCE],
+            DMongo.TIMESTAMP: rec[DMongo.TIMESTAMP],
+        }
+        db.delete_one("mining", {DMongo.OBJECT_ID: rec[DMongo.OBJECT_ID]})
+        db.insert_one("mining", new_rec)
+
+
+def migrate_block_found_events():
+    recs = db.find_many(col_name="tmp", filter={"doc_type": "block_found_event"})
+
+    for rec in recs:
+        new_rec = {
+            DMongo.DOC_TYPE: "block_found_event",
+            DMongo.TIMESTAMP: rec[DMongo.TIMESTAMP],
+            DMongo.INSTANCE: "Mini"
+        }
+        db.insert_one("mining", new_rec)
 
 
 def migrate_pool_hashrates():
@@ -46,6 +69,7 @@ def modify_pool_hashrates():
         db.update_one("mining", {DMongo.OBJECT_ID: rec[DMongo.OBJECT_ID]}, rec)
 
 
-modify_pool_hashrates()
+#modify_pool_hashrates()
 #migrate_pool_hashrates()
-# migrate_pool_hashrates()
+#migrate_block_found_events()
+modify_block_found_events_2()